refactor repository hierarchy
