+++ backend/app/core/databases/postgresdb.py
@@ -1,11 +1,11 @@
 from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine, async_sessionmaker
-from .config import db_url
-from .models import Base
+from backend.app.config import db_url
+from backend.app.models.models import Base
 
 
 engine = create_async_engine(
     url=str(db_url),
-    echo=True,
+    echo=False,
     future=True,
     )
