+++ backend/app/main.py
@@ -1,17 +1,19 @@
 import os
 from contextlib import asynccontextmanager
 
-from fastapi import FastAPI
+from fastapi import FastAPI, Depends
 from fastapi.middleware.cors import CORSMiddleware
 
-from backend.app.auth.routes import router as auth_router
-from backend.app.users.routes import router as users_router
-from backend.app.inn_service.routes import router as inn_service_router
-from backend.app.database import create_tables, Base
-from backend.app.enterprises.routes import router as enterprises_router
-from backend.app.reviews.routes import router as reviews_router
-from backend.app.mongodb.routes import router as mongodb_router
-from backend.app.points.routes import router as point_router
+from backend.app.api.v1.verify_routes import router as verify_router
+from backend.app.services.auth_bearer import JWTBearer
+from backend.app.api.v1.users_routes import router as users_router
+from backend.app.api.v1.inn_routes import router as inn_service_router
+from backend.app.core.databases.postgresdb import create_tables
+from backend.app.api.v1.enterprises_routes import router as enterprises_router
+from backend.app.api.v1.reviews_routes import router as reviews_router
+from backend.app.api.v1.mongo_routes import router as mongodb_router
+from backend.app.api.v1.points_routes import router as point_router
+from backend.app.api.v1.auth_routes import router as auth_router
 
 
 @asynccontextmanager
@@ -30,10 +32,12 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
+
+app.include_router(auth_router)
+app.include_router(verify_router)
 app.include_router(inn_service_router)
 app.include_router(users_router)
-app.include_router(auth_router)
 app.include_router(enterprises_router)
-app.include_router(reviews_router)
 app.include_router(point_router)
+app.include_router(reviews_router)
 app.include_router(mongodb_router)
