+++ backend/app/models/verify_models.py
@@ -1,6 +1,6 @@
 from sqlalchemy import String
 from sqlalchemy.orm import Mapped, mapped_column
-from backend.app.models import Base
+from backend.app.models.models import Base
 
 
 class Verification(Base):
